+++ lib/src/client_io.dart
@@ -58,7 +58,7 @@ class ClientIO extends ClientBase with ClientMixin {
       'x-sdk-name': 'Flutter',
       'x-sdk-platform': 'client',
       'x-sdk-language': 'flutter',
-      'x-sdk-version': '25.2.0',
+      'x-sdk-version': '25.3.0',
       'X-Appwrite-Response-Format': '1.9.5',
     };
 
@@ -94,6 +94,14 @@ class ClientIO extends ClientBase with ClientMixin {
   ClientIO setJWT(value) {
     config['jWT'] = value;
     addHeader('X-Appwrite-JWT', value);
+    return this;
+  }
+
+  /// The OAuth access token to authenticate with
+  @override
+  ClientIO setBearer(value) {
+    config['bearer'] = value;
+    addHeader('Authorization', 'Bearer $value');
     return this;
   }