+++ test/src/models/execution_list_test.dart
@@ -1,4 +1,5 @@
 import 'package:appwrite/models.dart';
+import 'package:appwrite/enums.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {